worker/requirements.txt
flask==3.0.0
google-cloud-firestore==2.13.1
gunicorn==21.2.0

